requirements.txt
discord.py>=2.3.0
openai>=1.0.0
aiohttp>=3.8.0
Pillow>=9.0.0
python-dotenv>=1.0.0
PyNaCl>=1.5.0
